# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Task Control Backend (Gateway)
  version: "1.0.0"
  description: >
    REST API, проходящее через API Gateway. Все ответы завернуты в объект с полем `success`.
servers:
  - url: http://localhost:8000
    description: Local via docker-compose
paths:
...
  /v1/users/register:
    post:
      tags: [Users]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "409":
          description: Email already used
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/users/login:
    post:
      tags: [Users]
      summary: Sign in and get JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/users/me:
    get:
      tags: [Users]
      summary: Current profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Users]
      summary: Update current profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/users:
    get:
      tags: [Users]
      summary: List users (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: email
          schema:
            type: string
        - in: query
          name: name
          schema:
            type: string
        - in: query
          name: role
          schema:
            type: string
      responses:
        "200":
          description: Paged list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/users/{id}:
    get:
      tags: [Users]
      summary: Get user by id (self or admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          required: true
          name: id
          schema:
            type: string
      responses:
        "200":
          description: User payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
        roles:
          type: array
          items:
            type: string
          description: Optional, defaults to ["user"]
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        roles:
          type: array
          items:
            type: string
          description: Allowed only for admin
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        roles:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/User'
    UserListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/User'
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            hasNext:
              type: boolean
            hasPrev:
              type: boolean
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string